backend/requirements.txt
fastapi
uvicorn
python-socketio
python-chess
python-multipart
python-dotenv
passlib[bcrypt]
PyJWT
sqlalchemy
